Finite element preconditioner works, but is actually not as strong as the finite difference one.
After much bug hunting, I became convinced that the inner part is actually
correct, but the Q1 finite element preconditioner is just not as good as the
finite difference preconditioner, at least applied naively.  I believe the root
of the problem is that we need to apply the mass matrix in order to get the
scaling correct.