Add plain C support, testing conditioning of perturbed matrices
Makefile: Allow a plain C version of PETSc to be used
stokes.C: Testing conditioning of perturbed matrices

Add option to filter matrix for testing effect of imposing
Dirichlet conditions by zeroing diagonal.  The option are only effective if
-pc_type 1 (Q1 finite element) preconditioning is chosen.  There are two new
options:

  -zeroN Int      number of rows and columns to eliminate
  -zeroV Scalar   value to place on the diagonal

We can test the conditioning of the velocity matrix with

  -vel_ksp_type preonly -vel_pc_type ksp -vel_ksp_ksp_rtol 1e-10 \
  -vel_ksp_ksp_gmres_restart 500 -vel_ksp_pc_type none \
  -vel_ksp_ksp_monitor_singular_value

and the conditioning of the Schur complement with

  -schur_pc_type none -svel_ksp_type preonly -svel_pc_type ksp \
  -svel_ksp_rtol 1e-10 -svel_ksp_pc_type ilu -schur_ksp_gmres_restart 500 \
  -schur_ksp_monitor_singular_value

Note that the divergence matrices are still applied to spectral order and
matrix-free.  Hence the procedure here is only a first-order approximation
to the effect on conditioning since we are not removing the appropriate rows
and columns from these matrix-free operators.